fixup! fuzz: seqnr_{aad,payload} populated via ConsumeIntegral<uint64_t>

+++ src/test/fuzz/crypto_chacha20_poly1305_aead.cpp
@@ -47,7 +47,7 @@ void test_one_input(const std::vector<uint8_t>& buffer)
             break;
         }
         case 3: {
-	    if (AdditionOverflow(seqnr_payload, 1)) {
+	    if (AdditionOverflow(seqnr_payload, (unsigned long)1)) {
 	        seqnr_payload = 1;
 	    } else {
 	        seqnr_payload += 1;
@@ -55,7 +55,7 @@ void test_one_input(const std::vector<uint8_t>& buffer)
             aad_pos += CHACHA20_POLY1305_AEAD_AAD_LEN;
             if (aad_pos + CHACHA20_POLY1305_AEAD_AAD_LEN > CHACHA20_ROUND_OUTPUT) {
                 aad_pos = 0;
-		if (AdditionOverflow(seqnr_aad, 1)) {
+		if (AdditionOverflow(seqnr_aad, (unsigned long)1)) {
 		    seqnr_aad = 1;
 		} else {
 		    seqnr_aad += 1;
